feat: fetch FileDef instances directly in render-store context
When prerendering a card that references files (e.g., images via
linksTo(FileDef)), the render store now bypasses the index and fetches
file bytes directly from the realm server, running extractAttributes
on the fly. This ensures subclass-specific attributes (width/height
for images, title/content for markdown) are available even when the
referenced file hasn't been indexed yet.

- Extract resolveFileDefCodeRef to shared runtime-common module
- Extract FileDefAttributesExtractor to shared host utility
- Add extractFileMetaDirectly render-store path in getFileMetaInstance
- Update test to exercise card-before-file indexing order
