+++ packages/realm-server/tests/card-endpoints-test.ts
@@ -4,7 +4,7 @@ import supertest from 'supertest';
 import { join, basename } from 'path';
 import type { Server } from 'http';
 import type { DirResult } from 'tmp';
-import { existsSync, readJSONSync, statSync } from 'fs-extra';
+import { existsSync, readJSONSync, statSync, writeFileSync } from 'fs-extra';
 import type {
   Realm,
   Relationship,
@@ -268,12 +268,25 @@ module(basename(__filename), function () {
         });
 
         test('includes FileDef resources for file links in included payload', async function (assert) {
-          let { testRealm: realm, request } = getRealmSetup();
+          let { testRealm: realm, request, dir: testDir } = getRealmSetup();
 
-          let writes = new Map<string, string | Uint8Array>([
-            [
-              'gallery.gts',
-              `
+          // Write image files directly to the filesystem so they are on disk
+          // but NOT yet in the index. This exercises the render-store's
+          // extractFileMetaDirectly path: when the card is prerendered, the
+          // images haven't been indexed yet, so getFileMetaInstance must fetch
+          // and extract attributes directly from the raw file bytes.
+          let realmDir = join(testDir.name, 'realm_server_1', 'test');
+          let pngBytes = makeMinimalPng();
+          writeFileSync(join(realmDir, 'hero.png'), pngBytes);
+          writeFileSync(join(realmDir, 'first.png'), pngBytes);
+          writeFileSync(join(realmDir, 'second.png'), pngBytes);
+
+          // Write module + card instance — card is indexed before images
+          await realm.writeMany(
+            new Map<string, string>([
+              [
+                'gallery.gts',
+                `
                 import { CardDef, field, linksTo, linksToMany } from "https://cardstack.com/base/card-api";
                 import { FileDef } from "https://cardstack.com/base/file-api";
 
@@ -282,44 +295,49 @@ module(basename(__filename), function () {
                   @field attachments = linksToMany(FileDef);
                 }
               `,
-            ],
-            [
-              'gallery.json',
-              JSON.stringify({
-                data: {
-                  attributes: {},
-                  relationships: {
-                    hero: {
-                      links: {
-                        self: './hero.png',
+              ],
+              [
+                'gallery.json',
+                JSON.stringify({
+                  data: {
+                    attributes: {},
+                    relationships: {
+                      hero: {
+                        links: {
+                          self: './hero.png',
+                        },
+                      },
+                      'attachments.0': {
+                        links: {
+                          self: './first.png',
+                        },
+                      },
+                      'attachments.1': {
+                        links: {
+                          self: './second.png',
+                        },
                       },
                     },
-                    'attachments.0': {
-                      links: {
-                        self: './first.png',
-                      },
-                    },
-                    'attachments.1': {
-                      links: {
-                        self: './second.png',
+                    meta: {
+                      adoptsFrom: {
+                        module: './gallery.gts',
+                        name: 'Gallery',
                       },
                     },
                   },
-                  meta: {
-                    adoptsFrom: {
-                      module: './gallery.gts',
-                      name: 'Gallery',
-                    },
-                  },
-                },
-              }),
-            ],
-            ['hero.png', makeMinimalPng()],
-            ['first.png', makeMinimalPng()],
-            ['second.png', makeMinimalPng()],
-          ]);
+                }),
+              ],
+            ]),
+          );
 
-          await realm.writeMany(writes);
+          // Now index the image files so they appear in loadLinks results
+          await realm.writeMany(
+            new Map<string, Uint8Array>([
+              ['hero.png', pngBytes],
+              ['first.png', pngBytes],
+              ['second.png', pngBytes],
+            ]),
+          );
 
           let response = await request
             .get('/gallery')
