+++ packages/runtime-common/index.ts
@@ -248,6 +248,7 @@ export * from './url';
 export * from './render-route-options';
 export * from './publishability';
 export * from './pr-manifest';
+export * from './file-def-code-ref';
 
 export const executableExtensions = ['.js', '.gjs', '.ts', '.gts'];
 export { createResponse } from './create-response';
